Fixes bug and format source code.
