+++ AMSExplorer/AMSClient/CredentialsEntryV3.cs
@@ -1,10 +1,7 @@
 ﻿using Microsoft.Azure.Management.Media.Models;
 using System;
-using System.Collections.Generic;
-using System.Linq;
 using System.Security.Cryptography;
 using System.Text;
-using System.Threading.Tasks;
 
 namespace AMSExplorer
 {
