+++ AMSExplorer/Forms-Live/LiveOutputInformation.cs
@@ -15,7 +15,6 @@
 //---------------------------------------------------------------------------------------------
 
 
-using Microsoft.Azure.Management.Media;
 using Microsoft.Azure.Management.Media.Models;
 using System;
 using System.Collections.Generic;
