+++ AMSExplorer/Forms-Login/EmbeddedBrowserFormUI.cs
@@ -1,18 +1,11 @@
-﻿using Microsoft.Identity.Client.Extensibility;
+﻿using Microsoft.Identity.Client;
+using Microsoft.Web.WebView2.Core;
 using System;
-using System.Collections.Generic;
-using System.ComponentModel;
-using System.Data;
-using System.Drawing;
 using System.Linq;
-using System.Text;
 using System.Threading;
 using System.Threading.Tasks;
 using System.Web;
 using System.Windows.Forms;
-using Microsoft.Identity.Client;
-using Microsoft.Web.WebView2.WinForms;
-using Microsoft.Web.WebView2.Core;
 
 namespace AMSExplorer.AMSLogin
 {
