+++ AMSExplorer/GridTransfer.cs
@@ -20,14 +20,13 @@ using System.ComponentModel;
 using System.Data;
 using System.Diagnostics;
 using System.Linq;
-using System.Runtime.CompilerServices;
 using System.Threading;
 using System.Threading.Tasks;
 using System.Windows.Forms;
 
 namespace AMSExplorer
 {
-    
+
     public partial class Mainform : Form
     {
         private static BindingList<TransferEntry> _MyListTransfer; // list of upload/download
