+++ AMSExplorer/JobInformation.cs
@@ -16,7 +16,6 @@
 
 
 using AMSExplorer.Utils.JobInfo;
-using Microsoft.Azure.Management.Media;
 using Microsoft.Azure.Management.Media.Models;
 using System;
 using System.Collections.Generic;
