+++ AMSExplorer/ListViews/ListViewTransforms.cs
@@ -93,7 +93,7 @@ namespace AMSExplorer
         }
         private async Task LoadTransformsAsync()
         {
-            
+
 
             _transforms = await _client.AMSclient.Transforms.ListAsync(_client.credentialsEntry.ResourceGroup, _client.credentialsEntry.AccountName);
 
@@ -124,7 +124,7 @@ namespace AMSExplorer
                 {
                     try
                     {
-                        
+
                         await _client.AMSclient.Transforms.DeleteAsync(_client.credentialsEntry.ResourceGroup, _client.credentialsEntry.AccountName, transform.Name);
                     }
                     catch
