+++ AMSExplorer/Models/Live/OperationType.cs
@@ -28,5 +28,5 @@ namespace AMSExplorer
         ResetAsset,
         Reset
     }
-          
+
 }
