+++ AMSExplorer/Rest/TransformRestObject.cs
@@ -16,7 +16,6 @@
 
 using Newtonsoft.Json;
 using System.Collections.Generic;
-using System.Threading.Tasks;
 
 namespace AMSExplorer.Rest
 {
