+++ AMSExplorer/Splash.cs
@@ -15,7 +15,6 @@
 //---------------------------------------------------------------------------------------------
 
 using System;
-using System.Collections.Generic;
 using System.Reflection;
 using System.Windows.Forms;
