+++ AMSExplorer/Utils/TransformInfo/TransformTools.cs
@@ -1,12 +1,8 @@
 ﻿using AMSExplorer.Rest;
 using Microsoft.Azure.Management.Media.Models;
 using Newtonsoft.Json;
-using System;
-using System.Collections.Generic;
 using System.Linq;
 using System.Text;
-using System.Threading.Tasks;
-using System.Windows.Forms;
 
 namespace AMSExplorer.Utils.TransformInfo
 {
